Remove buggy special case from Cmp, use Copy not Set in tests
Fixes: #15

Former-commit-id: 3bae9bd699eba979f677c39ee7b99bc535e140be [formerly 0824ccd7dc17c3db54f2ac23fe42fdedce0b69c6]
Former-commit-id: 2f5a32fab7d9e10c957fe42e8d8dc087ba9b567a

+++ decimal_perf_test.go
@@ -45,7 +45,7 @@ var zeroPool = sync.Pool{
 }
 
 var getInt = func(x int64) *big.Int {
-	return intPool.Get().(*big.Int).SetInt(x)
+	return intPool.Get().(*big.Int).SetInt64(x)
 }
 
 var putInt = func(x *big.Int) {
